requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-docx==0.8.11
PyPDF2==3.0.1
openpyxl==3.1.2
pandas==2.1.1
requests==2.31.0
python-multipart==0.0.6
Pillow==10.0.1
python-dotenv==1.0.0
